cleanup for production
